phc2sys: break out pmc code into pmc_common.c
Move the code for sending various messages to ptp4l via pmc to a common
translation module, outside of phc2sys. This makes it available to other
programs that want to subscribe to port state change events too, such as
ts2phc.

This creates a smaller structure within phc2sys_private, which embeds
all properties related to the PMC. This structure is called "pmc_node",
which is somewhat reminiscent of the old name of phc2sys_private (struct
node). But the advantage is that struct pmc_node can be reused by other
modules.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>