ts2phc: create a private data structure
Eliminate the use of global variables in the ts2phc program by
introducing one data structure that incorporates them. This might make
the code more understandable to people coming from a kernel background,
since it resembles the type of data organization used there. It is also
now closer to the data organization of phc2sys, a similar program in
both purpose and implementation.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>