pmc_common: fix potential memory leak in run_pmc_events()
The convention in all other parts of the code that call run_pmc() is to
free the management PTP message if an error code was returned, or pass
the message to the caller otherwise.

run_pmc_events() wasn't doing that, and was leaking a reference to the
message, while also discarding the return code from run_pmc(). Fix that.

Signed-off-by: Vladimir Oltean <olteanv@gmail.com>